requirements.txt
flask
pillow
qrcode